+++ src/fields/Table.php
@@ -24,6 +24,7 @@ use craft\web\assets\tablesettings\TableSettingsAsset;
 use craft\web\assets\timepicker\TimepickerAsset;
 use GraphQL\Type\Definition\InputObjectType;
 use GraphQL\Type\Definition\Type;
+use LitEmoji\LitEmoji;
 use yii\db\Schema;
 use yii\validators\EmailValidator;
 
@@ -419,7 +420,13 @@ class Table extends Field
         foreach ($value as $row) {
             $serializedRow = [];
             foreach (array_keys($this->columns) as $colId) {
-                $serializedRow[$colId] = parent::serializeValue($row[$colId] ?? null);
+                $value = $row[$colId];
+
+                if (is_string($value) && in_array($this->columns[$colId]['type'], ['singleline', 'multiline'], true)) {
+                    $value = LitEmoji::unicodeToShortcode($value);
+                }
+
+                $serializedRow[$colId] = parent::serializeValue($value ?? null);
             }
             $serialized[] = $serializedRow;
         }
@@ -501,6 +508,12 @@ class Table extends Field
 
                 return new ColorData($value);
 
+            case 'multiline':
+            case 'singleline':
+                if ($value !== null) {
+                    $value = LitEmoji::shortcodeToUnicode($value);
+                    return trim(preg_replace('/\R/u', "\n", $value));
+                }
             case 'date':
             case 'time':
                 return DateTimeHelper::toDateTime($value) ?: null;